requirements.txt
mysqlclient
djangorestframework
django
drf-nested-routers
django-jazzmin
djoser
djangorestframework-simplejwt
pillow
django-cors-headers
